refactored caseDataAggregator now a method in DataTools
